+++ app/src/main/java/com/mobile/tugasrancangmoka/utils/UriToFileUtil.kt
@@ -1,0 +1,24 @@
+package com.mobile.tugasrancangmoka.utils // sesuaikan dengan lokasi package-mu
+
+import android.content.Context
+import android.net.Uri
+import java.io.File
+import java.io.FileOutputStream
+import java.io.InputStream
+
+object UriToFileUtil {
+    fun getFileFromUri(context: Context, uri: Uri): File {
+        val contentResolver = context.contentResolver
+        val tempFile = File.createTempFile("upload_", ".jpg", context.cacheDir)
+
+        val inputStream: InputStream? = contentResolver.openInputStream(uri)
+        val outputStream = FileOutputStream(tempFile)
+
+        inputStream?.use { input ->
+            outputStream.use { output ->
+                input.copyTo(output)
+            }
+        }
+        return tempFile
+    }
+}